fixes and display navigation for mobile
